SpiQ

Documentation for SpQEphysTools.

List of functions:

extract_peaks(xf::Array{Float32,1}, thr::Float32, dpre::Float32, dpost::Float32, ref::Float32, event::Int64, srate::Float32)

prepare_bandpass(lowcut::Float32, highcut::Float32, fs::Float32)::ZPG

bandpass(data::Array{Float32,1}, filt::ZPG)

prepare_lowpass(cutoff::Float32, fs::Float32)

lowpassanddec(data::Array{Float32,1}, filt::ZPG, rate::Float32, fs::Float32)

allocateFloat32vector(numelements::Int64)

meminfo_julia()

preproc_chan(fname::String, chan::Int, datasetname::String)

parsetomlfiles(OUTPUT)::settings

tideup_output(OUTPUT)

report(text::String, filename::String)

active_el(filename::String, threshold::Float64)::Int

extract_bursts(pathname::String, nActiveEl::Int, ref::Float64)::Int

plot_raster(pathname::String, fraction::Float64=1.0, title::Bool=false)::Int

plot_frequencies(pathname::String, fraction::Float64=1.0, title::Bool=false)::Int

SpQEphysTools.allocate_Float32vector

SpQEphysTools.bandpass

SpQEphysTools.extract_bursts

SpQEphysTools.extract_peaks

SpQEphysTools.plot_frequencies

SpQEphysTools.plot_raster

SpQEphysTools.prepare_bandpass

SpQEphysTools.prepare_lowpass

SpQEphysTools.preproc_chan

SpQEphysTools.allocate_Float32vector — Method

allocate_Float32vector(num_elements::Int64)

Allocate a Float32 vector of a given size, handling OutOfMemoryError exceptions

SpQEphysTools.bandpass — Method

bandpass(data::Array{Float32,1}, lowcut::Float32, highcut::Float32, fs::Float32)

This function applies a bandpass filter to the input data. It receives the data (as a 1D array of Float32),
and the filter parameters previously defined by prepare_bandpass. It returns the filtered data.

SpQEphysTools.extract_bursts — Method

extract_bursts(pathname::String, nActiveEl::Int)::Int

Detect the bursts as in Van Pelt et al., 2004; IEEE Trans. Biomed. Eng., 51(11):2051-62. Alternatively, it uses a fixed threshold to detect the bursts, ignoring combined effect of an increase in the firing rate and an increase in the number of channels simultaneously active.

Arguments

pathname::String : path to the folder containing the spike times and the configuration files (the pre-processed *.dat folder)

nActiveEl::Int : number of active electrodes

Author(s)

Michele Giugliano - michele.giugliano@unimore.it

SpQEphysTools.extract_peaks — Method

extract_peaks(xf::Array{Float32,1}, thr::Float32, dpre::Float32, dpost::Float32, ref::Float32, event::Int64, srate::Float32)

Performs a threshold crossing detection on a given signal (xf), and extracts the peaks of the signal in terms of their
index/time of occurrence and "polarity" (i.e., sign of the peak). The function is designed to work with a single channel
signal, and it is used in the context of spike detection in electrophysiological recordings.

Arguments

It requires the following parameters:
- `xf::Array{Float32,1}` : the signal to be analyzed.
- `thr::Float32` : the threshold value for the detection.
- `dpre::Float32` : the duration of the pre-event window (in ms), for storing the event waveform.
- `dpost::Float32` : the duration of the post-event window (in ms), for storing the event waveform.
- `ref::Float32` : the refractory period (in ms), to avoid multiple detections of the same event.
- `event::Int64` : the type of event to be detected (1 for positive, -1 for negative, 0 for both).
- `srate::Float32` : the sampling rate of the signal (in Hz).

Author(s)

Michele Giugliano - michele.giugliano@unimore.it

SpQEphysTools.plot_frequencies — Function

plot_frequencies(pathname::String, fraction::Float64)

Plot the distribution of firing rates, for each fraction of the total recording time. The function generates one or more PDF figures, storing them in the folder figs within the input pathname. If fraction is set to 1.0, then only one figure is generated, containing the entire recording time. Otherwise, multiple figures are generated, each containing a fraction of the total recording time.

Arguments

pathname::String : path to the folder containing the spike times and the configuration files (the pre-processed *.dat folder)

fraction::Float64 : fraction of the total recording time (e.g. 0.1; must be in the range (0, 1)) (default: 1.0)

title::Bool : if true, the title of the plot is the name of the folder (default: false)

Author(s)

Michele Giugliano - michele.giugliano@unimore.it

SpQEphysTools.plot_raster — Function

plot_raster(pathname::String, fraction::Float64)

Plot the raster plot of the spike times, for each fraction of the total recording time. The function generates one or more PDF figures, storing them in the folder figs within the input pathname. If fraction is set to 1.0, then only one figure is generated, containing the entire recording time. Otherwise, multiple figures are generated, each containing a fraction of the total recording time.

Arguments

pathname::String : path to the folder containing the spike times and the configuration files (the pre-processed *.dat folder)

fraction::Float64 : fraction of the total recording time (e.g. 0.1; must be in the range (0, 1)) (default: 1.0)

title::Bool : if true, the title of the plot is the name of the folder (default: false)

Author(s)

Michele Giugliano - michele.giugliano@unimore.it

SpQEphysTools.prepare_bandpass — Method

prepare_bandpass(lowcut::Float32, highcut::Float32, fs::Float32)

This function prepares (once for all) the bandpass filter to be used for spike detection/sorting.
It returns the filter object. It receives the low and high cutoff frequencies (in Hz) and the sampling rate (in Hz).
Its purpose is to avoid redefining the filter at each call of the bandpass function.
Its output is used as input to the bandpass function.

SpQEphysTools.prepare_lowpass — Method

prepare_lowpass(cutoff::Float32, fs::Float32)

This function prepares (once for all) the lowpass filter to be used for LFP extraction.
It returns the filter object. It receives the low cutoff frequency (in Hz) and the sampling rate (in Hz).
Its purpose is to avoid redefining the filter at each call of the lowpass function.
Its output is used as input to the lowpass function.

SpQEphysTools.preproc_chan — Method

preproc_chan(fname, chan, datasetname)

TBW